rag-chatbot/backend/requirements.txt
asyncpg==0.29.0
pydantic==2.9.2
uvicorn==0.30.6
SQLAlchemy==2.0.35
pgvector==0.3.3
pymupdf4llm==0.0.17
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
fastapi==0.115.0
langchain-core==0.3.6
langchain-experimental==0.3.2
langchain-ollama==0.2.0
tqdm==4.66.5